mocks moved to a file, code clean up
