+++ one-shot-device/src/devices/device-registration.service.spec.ts
@@ -6,31 +6,14 @@ import { MongooseModule } from '@nestjs/mongoose';
 import { DeviceRegistration, DeviceRegistrationDocument, DeviceRegistrationSchema } from './schemas/device-registration.schema';
 import { CreateDeviceRegistrationDto as dto } from './dto/create-device-registration.dto';
 import { Test, TestingModule } from '@nestjs/testing';
-import { Model } from 'mongoose';
 import { MongoMemoryServer } from 'mongodb-memory-server';
+import { mockDeviceRegistration } from './mocks';
 
 jest.setTimeout(40000);
 describe('DeviceRegistrationController', () => {
 	let deviceRegistrationController: DeviceRegistrationController;
 	let deviceRegistrationService: DeviceRegistrationService;
-	let deviceRegistrationModel: Model<DeviceRegistrationDocument>;
 	let module: TestingModule;
-
-	const mockData: dto | any = {
-		identityKeys: {
-			id: 'did:iota:5qmWjFWcqE3BNTB9KNCBH6reXEgig4gFUXiPENywB3wo',
-			key: {
-				type: 'ed25519',
-				public: 'BKU1cLqakKrN9g4ridQ6z5NyHqA7vZLLqmAygwDrSeex',
-				secret: '9EXX7aqBKGpBBDvy7ND65PvHe2DHwZrq7ZLorzE8ZDvv',
-				encoding: 'base58'
-			}
-		},
-		channelSeed: 'jldirikybrxczxlhhswzikqhsafjdzejjacoqaymzmoffdrwrzmytolrwuyhwoweybnzofew',
-		channelId: 'f48875646434e9b12019d2290bd74f0f4eae8393ada3b503202dfc713f0323070000000000000000:3cce98eb1742468ff35fde6b',
-		nonce: '1b0e4a49-3a23-4e7e-99f4-97fda845ff02'
-	};
-
 	afterEach(() => {
 		module.close();
 	});
@@ -61,5 +44,17 @@ describe('DeviceRegistrationController', () => {
 
 	it('deviceRegistrationService should be defined', () => {
 		expect(deviceRegistrationService).toBeDefined();
+	});
+
+	it('should create channel', async () => {
+		// const channel = jest.spyOn(deviceRegistrationService, 'createChannelAndIdentity');
+	});
+
+	it('should create Identity', async () => {
+		// const identity = jest.spyOn(deviceRegistrationService, 'createChannelAndIdentity');
+	});
+
+	it('should authenticate the device', async () => {
+		// const identity = jest.spyOn(deviceRegistrationService, 'createChannelAndIdentity');
 	});
 }); // describe block end here
